requirements.txt
opensearch-py
sentence-transformers
google-genai
